escola/requirements.txt
django
djangorestframework
python-gcm
gunicorn
psycopg2
django-filter